moving default testnet config
